Fix missed casts to double

+++ src/main/java/net/dmulloy2/ultimatearena/types/ArenaPlayer.java
@@ -317,12 +317,12 @@ public class ArenaPlayer
 	 */
 	public double getKDR()
 	{
-		double k = kills;
+		double k = (double) kills;
 		if (deaths == 0)
 			return k;
 
-		double d = deaths;
-		return (k / d);
+		double d = (double) deaths;
+		return k / d;
 	}
 
 	private long deathTime;
